Initial commit.
This is a pseudo-fork from Tool 1 (version 5.3.0 RC1 as of July,
21st 2025, commit #b73557c). This snapshot has nothing to do with
Tool 3 yet and is just another copy from Tool 1. It will serve as
the foundation of Tool 3. Functionalities will be integrated into
this code base one by one.

The original copy of Tool 3 version 3 (English) is available as
a tagged release on GitHub. This release corresponds to the current
production version of Tool3v3En deployed to shinyapps.io. It is not
a new release per se (only a back-up of an important past version).