Initial commit on read_data().
This function is the main reading mechanism for the external
measurement dataset. It currently support CSV, XLS, and XLSX
files. The objective is to encapsulate the reading and importing
logic into its own component (outside of the server code). It
sets important values that will have to be conveyed to the user
in a future FAQ section.