New observer for input btn_clear.
The objective is to completely reset the uploaded file whenever
this button is pressed by the user. The file is removed from the
server for safety reasons, the state of the input dataset is
reset, choices of input data_variables are removed, and inputs
data_variable and btn_submit are disabled until a new valid file
is uploaded.