Allow subsets of measurements in the Descriptive Statistics Panel module.
This follows previous commits. This module must be used in the
context of single-category risk analyses, similarly to how it
is used in Tool 1. However, this requires further logic to
intelligently create subsets of measurements from data_sample().
This is accomplished by select.cat.formatted(). The latter is
called by the data_sample_imputed() reactive object (minimizing
number of otherwise required changes elsewhere).